requirements-macos.txt
# Core ML/AI Dependencies
torch==2.11.0
torchvision==0.26.0
tensorflow
onnxruntime==1.23.2

numpy==2.1.3

# HTTP and SSL dependencies
certifi

# Computer Vision
opencv-python==4.11.0.86
rawpy==0.26.1

# Data Processing
pandas==2.2.2

# GUI Framework
PyQt6==6.10.2

# Utilities
PyExifTool==0.5.6
requests==2.33.0
pywebview==6.1

# installer
pyinstaller==6.18.0

# EXIF reading
exifread
pillow==12.1.1